refactor: where stuff lives
